# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Registro Epidemiologico de Doenças Raras",
        "description": "Aplicação REST que consiste em um registro epidemiológico para pacientes com doenças raras a\n    servir aplicações que podem ser outros sistemas ou uma interface de usuário.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "localhost:8000/registro_epidemiologico/api",
            "description": ""
        },
        {
            "url": "localhost:8000/registro_epidemiologico/"
        }
    ],
    "paths": {
...
        "/auth/login": {
            "post": {
                "tags": [
                    "autenticacao"
                ],
                "summary": "Registra o login do usuário",
                "description": "Faz o login do autenticado para que ele possa ter acesso às funcionalidades que precisam de autenticação. Retorna os dados de acesso.",
                "operationId": "login",
                "requestBody": {
                    "description": "Registra o login do usuário",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "example": "juliana@email.com"
                                    },
                                    "password": {
                                        "type": "integer",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Usuário não autorizado"
                    },
                    "400": {
                        "description": "Erros de validação dos campos passados no request."
                    },
                    "200": {
                        "description": "Operação realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.\n     *                  eyJpc3MiOiJodHRwOlwvXC8xMjcuMC4wLjE6ODAwMFwvYXBpXC9hdXRoXC9sb2dpbiIsImlhdCI6MTYxNTMwODQ3OCwi\n     *                  ZXhwIjoxNjE1Mzk0ODc4LCJuYmYiOjE2MTUzMDg0NzgsImp0aSI6Ikdwb3Y4UDZhRWszNnJwT24iLCJzdWIiOjMsInBy\n     *                  diI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.dQ6TIYmYo25Neveic_BB6-xEno-2BQ1n2pca0oJfyDs"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "token_validity": {
                                            "type": "integer",
                                            "example": "86400"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "autenticacao"
                ],
                "summary": "Cadastra novo usuário",
                "description": "Realiza o cadastro do usuário e retorna os dados cadastrados.",
                "operationId": "register",
                "requestBody": {
                    "description": "Cadastra o usuário",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Juliana Alves"
                                    },
                                    "email": {
                                        "type": "email",
                                        "example": "juliana@email.com"
                                    },
                                    "password": {
                                        "type": "integer",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Usuário não autorizado"
                    },
                    "400": {
                        "description": "Erros de validação dos campos passados no request."
                    },
                    "200": {
                        "description": "Operação realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "nome": {
                                            "type": "string",
                                            "example": "Juliana Alves"
                                        },
                                        "email": {
                                            "type": "email",
                                            "example": "juliana@email.com"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "autenticacao"
                ],
                "summary": "Faz o logout do usuário",
                "description": "Realiza o logout do usuário.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Operação realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário deslogado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/profle": {
            "get": {
                "tags": [
                    "autenticacao"
                ],
                "summary": "Retorna usuário.",
                "description": "Retorna os dados do susuário logado.",
                "operationId": "profile",
                "responses": {
                    "200": {
                        "description": "Operação realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "nome": {
                                            "type": "string",
                                            "example": "Juliana Alves"
                                        },
                                        "email": {
                                            "type": "email",
                                            "example": "juliana@email.com"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "autenticacao"
                ],
                "summary": "Atualiza token",
                "description": "Atualiza os dados de acesso do usuário logado. Retorna os dados de acesso.",
                "operationId": "refresh",
                "responses": {
                    "401": {
                        "description": "Usuário não autorizado"
                    },
                    "200": {
                        "description": "Operação realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.\n     *                  eyJpc3MiOiJodHRwOlwvXC8xMjcuMC4wLjE6ODAwMFwvYXBpXC9hdXRoXC9sb2dpbiIsImlhdCI6MTYxNTMwODQ3OCwi\n     *                  ZXhwIjoxNjE1Mzk0ODc4LCJuYmYiOjE2MTUzMDg0NzgsImp0aSI6Ikdwb3Y4UDZhRWszNnJwT24iLCJzdWIiOjMsInBy\n     *                  diI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.dQ6TIYmYo25Neveic_BB6-xEno-2BQ1n2pca0oJfyDs"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "token_validity": {
                                            "type": "integer",
                                            "example": "86400"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Apenas usuários autenticados podem realizar esta operação",
                "name": "token",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}